+++ src/main.rs
@@ -573,3 +573,4 @@ impl_from_error!(jiff::Error);
 impl_from_error!(Utf8Error);
 impl_from_error!(btoi::ParseIntegerError);
 impl_from_error!(fast_float::Error);
+impl_from_error!(std::num::ParseIntError);