+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Adding `xan explode -e/--evaluate, -f/--evaluate-file, --pad & -k/--keep`.
 * `xan to npy` is now able to stream.
 * Adding `xan parallel top` & `xan top -p/--parallel, -t/--threads`.
+* Adding `xan network edgelist --range`.
 
 *Fixes*